refactor extradata codec
