+++ core/capabilities/ccip/ccipevm/extradatadecoder.go
@@ -3,8 +3,51 @@ package ccipevm
 import (
 	"fmt"
 
+	chainsel "github.com/smartcontractkit/chain-selectors"
+
 	cciptypes "github.com/smartcontractkit/chainlink-ccip/pkg/types/ccipocr3"
 )
+
+type ExtraDataCodec struct{}
+
+func NewExtraDataCodec() ExtraDataCodec {
+	return ExtraDataCodec{}
+}
+
+func (c ExtraDataCodec) DecodeExtraArgs(extraArgs cciptypes.Bytes, sourceChainSelector cciptypes.ChainSelector) (map[string]any, error) {
+	if len(extraArgs) == 0 {
+		// return empty map if extraArgs is empty
+		return nil, nil
+	}
+
+	family, err := chainsel.GetSelectorFamily(uint64(sourceChainSelector))
+	if err != nil {
+		return nil, fmt.Errorf("failed to get chain family for selector %d: %w", sourceChainSelector, err)
+	}
+
+	if family != chainsel.FamilyEVM {
+		return nil, fmt.Errorf("chain selector family is %d, expect %s", sourceChainSelector, chainsel.FamilyEVM)
+	}
+	return DecodeExtraArgsToMap(extraArgs)
+}
+
+func (c ExtraDataCodec) DecodeTokenAmountDestExecData(destExecData cciptypes.Bytes, sourceChainSelector cciptypes.ChainSelector) (map[string]any, error) {
+	if len(destExecData) == 0 {
+		// return empty map if destExecData is empty
+		return nil, nil
+	}
+
+	family, err := chainsel.GetSelectorFamily(uint64(sourceChainSelector))
+	if err != nil {
+		return nil, fmt.Errorf("failed to get chain family for selector %d: %w", sourceChainSelector, err)
+	}
+
+	if family != chainsel.FamilyEVM {
+		return nil, fmt.Errorf("chain selector family is %d, expect %s", sourceChainSelector, chainsel.FamilyEVM)
+	}
+
+	return DecodeDestExecDataToMap(destExecData)
+}
 
 func DecodeDestExecDataToMap(DestExecData cciptypes.Bytes) (map[string]interface{}, error) {
 	destGasAmount, err := abiDecodeUint32(DestExecData)
